Changes made:
Created a switcher function that changed the contents of the
CourseInfoBox pop up to include content for cores, electives, guided
electives, and prescribed electives. Will need to change json pixel data
to specify which electives are prescribed/guided per major. The url that
redirects the user generateMajorURL(major: string) dynamically creates
the url with the user selected major and catalog year.